Extract a distribution represented by samples

The distribution can be extracted as:

a probability density function (type = "continuous"), cf. extract_pdf().

a probability mass function (type = "discrete"), cf. extract_pmf().

a series of equal-tailed confidence/credible intervals (type = "eti"), cf. extract_ci().

a series of highest density confidence/credible intervals (type = "hdi"), cf. extract_ci().

extract_distribution(
  object,
  parName = "",
  type = c("continuous", "discrete", "eti", "hdi"),
  transform = identity,
  ...
)

Arguments

object
Object specifying the distribution as samples: can be a Stanfit object, a matrix (columns represents parameters, rows samples) or a vector.

parName
Name of the parameter to extract.

type
Indicates how the distribution is summarised.

transform
Function to apply to the samples.

...
Arguments to pass to extract_pmf(), extract_pdf() or extract_ci() depending on type.

Value

Dataframe. The columns depends on the method that is used (see specific function for details).

Alternative

This function can notably be used to prepare the data for plotting fan charts when type = "eti" or "hdi". In that case, the ggdist package offers an alternative with ggdist::stat_lineribbon().

See also

Examples

extract_distribution(runif(1e2), type = "continuous", support = c(0, 1))
#>           Value   Density Index Variable
#> 1   0.000000000 0.4787754    NA         
#> 2   0.007874016 0.5069943    NA         
#> 3   0.015748031 0.5351773    NA         
#> 4   0.023622047 0.5632345    NA         
#> 5   0.031496063 0.5910805    NA         
#> 6   0.039370079 0.6186383    NA         
#> 7   0.047244094 0.6458554    NA         
#> 8   0.055118110 0.6726822    NA         
#> 9   0.062992126 0.6990711    NA         
#> 10  0.070866142 0.7249745    NA         
#> 11  0.078740157 0.7503786    NA         
#> 12  0.086614173 0.7752559    NA         
#> 13  0.094488189 0.7995760    NA         
#> 14  0.102362205 0.8232878    NA         
#> 15  0.110236220 0.8463822    NA         
#> 16  0.118110236 0.8688213    NA         
#> 17  0.125984252 0.8905607    NA         
#> 18  0.133858268 0.9115094    NA         
#> 19  0.141732283 0.9316421    NA         
#> 20  0.149606299 0.9508923    NA         
#> 21  0.157480315 0.9691878    NA         
#> 22  0.165354331 0.9864057    NA         
#> 23  0.173228346 1.0025015    NA         
#> 24  0.181102362 1.0174050    NA         
#> 25  0.188976378 1.0310437    NA         
#> 26  0.196850394 1.0433026    NA         
#> 27  0.204724409 1.0541557    NA         
#> 28  0.212598425 1.0635808    NA         
#> 29  0.220472441 1.0715553    NA         
#> 30  0.228346457 1.0780333    NA         
#> 31  0.236220472 1.0830504    NA         
#> 32  0.244094488 1.0866770    NA         
#> 33  0.251968504 1.0889787    NA         
#> 34  0.259842520 1.0900173    NA         
#> 35  0.267716535 1.0899082    NA         
#> 36  0.275590551 1.0888136    NA         
#> 37  0.283464567 1.0868814    NA         
#> 38  0.291338583 1.0842636    NA         
#> 39  0.299212598 1.0811337    NA         
#> 40  0.307086614 1.0776882    NA         
#> 41  0.314960630 1.0741032    NA         
#> 42  0.322834646 1.0705522    NA         
#> 43  0.330708661 1.0672118    NA         
#> 44  0.338582677 1.0642278    NA         
#> 45  0.346456693 1.0617309    NA         
#> 46  0.354330709 1.0598344    NA         
#> 47  0.362204724 1.0586563    NA         
#> 48  0.370078740 1.0582286    NA         
#> 49  0.377952756 1.0585846    NA         
#> 50  0.385826772 1.0597314    NA         
#> 51  0.393700787 1.0616845    NA         
#> 52  0.401574803 1.0643627    NA         
#> 53  0.409448819 1.0676966    NA         
#> 54  0.417322835 1.0715993    NA         
#> 55  0.425196850 1.0759815    NA         
#> 56  0.433070866 1.0807063    NA         
#> 57  0.440944882 1.0856441    NA         
#> 58  0.448818898 1.0906651    NA         
#> 59  0.456692913 1.0956328    NA         
#> 60  0.464566929 1.1004136    NA         
#> 61  0.472440945 1.1048880    NA         
#> 62  0.480314961 1.1089455    NA         
#> 63  0.488188976 1.1124708    NA         
#> 64  0.496062992 1.1153782    NA         
#> 65  0.503937008 1.1176115    NA         
#> 66  0.511811024 1.1191191    NA         
#> 67  0.519685039 1.1198493    NA         
#> 68  0.527559055 1.1197772    NA         
#> 69  0.535433071 1.1189210    NA         
#> 70  0.543307087 1.1172909    NA         
#> 71  0.551181102 1.1148982    NA         
#> 72  0.559055118 1.1117634    NA         
#> 73  0.566929134 1.1079524    NA         
#> 74  0.574803150 1.1035111    NA         
#> 75  0.582677165 1.0984850    NA         
#> 76  0.590551181 1.0929118    NA         
#> 77  0.598425197 1.0868668    NA         
#> 78  0.606299213 1.0804011    NA         
#> 79  0.614173228 1.0735635    NA         
#> 80  0.622047244 1.0663884    NA         
#> 81  0.629921260 1.0589366    NA         
#> 82  0.637795276 1.0512490    NA         
#> 83  0.645669291 1.0433638    NA         
#> 84  0.653543307 1.0353116    NA         
#> 85  0.661417323 1.0271344    NA         
#> 86  0.669291339 1.0188661    NA         
#> 87  0.677165354 1.0105387    NA         
#> 88  0.685039370 1.0021840    NA         
#> 89  0.692913386 0.9938358    NA         
#> 90  0.700787402 0.9855249    NA         
#> 91  0.708661417 0.9772808    NA         
#> 92  0.716535433 0.9691349    NA         
#> 93  0.724409449 0.9611115    NA         
#> 94  0.732283465 0.9532286    NA         
#> 95  0.740157480 0.9454995    NA         
#> 96  0.748031496 0.9379337    NA         
#> 97  0.755905512 0.9305258    NA         
#> 98  0.763779528 0.9232551    NA         
#> 99  0.771653543 0.9160918    NA         
#> 100 0.779527559 0.9089918    NA         
#> 101 0.787401575 0.9018920    NA         
#> 102 0.795275591 0.8947147    NA         
#> 103 0.803149606 0.8873698    NA         
#> 104 0.811023622 0.8797485    NA         
#> 105 0.818897638 0.8717195    NA         
#> 106 0.826771654 0.8631671    NA         
#> 107 0.834645669 0.8539598    NA         
#> 108 0.842519685 0.8439575    NA         
#> 109 0.850393701 0.8329877    NA         
#> 110 0.858267717 0.8209582    NA         
#> 111 0.866141732 0.8077529    NA         
#> 112 0.874015748 0.7932698    NA         
#> 113 0.881889764 0.7773557    NA         
#> 114 0.889763780 0.7600137    NA         
#> 115 0.897637795 0.7412084    NA         
#> 116 0.905511811 0.7209309    NA         
#> 117 0.913385827 0.6991401    NA         
#> 118 0.921259843 0.6759400    NA         
#> 119 0.929133858 0.6514177    NA         
#> 120 0.937007874 0.6256776    NA         
#> 121 0.944881890 0.5988141    NA         
#> 122 0.952755906 0.5710126    NA         
#> 123 0.960629921 0.5424571    NA         
#> 124 0.968503937 0.5133319    NA         
#> 125 0.976377953 0.4838258    NA         
#> 126 0.984251969 0.4541553    NA         
#> 127 0.992125984 0.4245283    NA         
#> 128 1.000000000 0.3951427    NA