Extract a distribution represented by samples

The distribution can be extracted as:

a probability density function (type = "continuous"), cf. extract_pdf().

a probability mass function (type = "discrete"), cf. extract_pmf().

a series of equal-tailed confidence/credible intervals (type = "eti"), cf. extract_ci().

a series of highest density confidence/credible intervals (type = "hdi"), cf. extract_ci().

extract_distribution(
  object,
  parName = "",
  type = c("continuous", "discrete", "eti", "hdi"),
  transform = identity,
  ...
)

Arguments

object

Object specifying the distribution as samples: can be a Stanfit object, a matrix (columns represents parameters, rows samples) or a vector.

parName

Name of the parameter to extract.

type

Indicates how the distribution is summarised.

transform

Function to apply to the samples.

...

Arguments to pass to extract_pmf(), extract_pdf() or extract_ci() depending on type.

Value

Dataframe. The columns depends on the method that is used (see specific function for details).

Alternative

This function can notably be used to prepare the data for plotting fan charts when type = "eti" or "hdi". In that case, the ggdist package offers an alternative with ggdist::stat_lineribbon().

See also

Examples

extract_distribution(runif(1e2), type = "continuous", support = c(0, 1))
#>           Value   Density Index Variable
#> 1   0.000000000 0.6340619    NA         
#> 2   0.007874016 0.6626269    NA         
#> 3   0.015748031 0.6898039    NA         
#> 4   0.023622047 0.7154130    NA         
#> 5   0.031496063 0.7392737    NA         
#> 6   0.039370079 0.7612272    NA         
#> 7   0.047244094 0.7811437    NA         
#> 8   0.055118110 0.7989249    NA         
#> 9   0.062992126 0.8145064    NA         
#> 10  0.070866142 0.8278210    NA         
#> 11  0.078740157 0.8388853    NA         
#> 12  0.086614173 0.8477722    NA         
#> 13  0.094488189 0.8545595    NA         
#> 14  0.102362205 0.8593565    NA         
#> 15  0.110236220 0.8623014    NA         
#> 16  0.118110236 0.8635579    NA         
#> 17  0.125984252 0.8633025    NA         
#> 18  0.133858268 0.8617232    NA         
#> 19  0.141732283 0.8590697    NA         
#> 20  0.149606299 0.8555639    NA         
#> 21  0.157480315 0.8514295    NA         
#> 22  0.165354331 0.8468870    NA         
#> 23  0.173228346 0.8421492    NA         
#> 24  0.181102362 0.8374172    NA         
#> 25  0.188976378 0.8328851    NA         
#> 26  0.196850394 0.8287184    NA         
#> 27  0.204724409 0.8250586    NA         
#> 28  0.212598425 0.8220237    NA         
#> 29  0.220472441 0.8197067    NA         
#> 30  0.228346457 0.8181747    NA         
#> 31  0.236220472 0.8174694    NA         
#> 32  0.244094488 0.8176214    NA         
#> 33  0.251968504 0.8186210    NA         
#> 34  0.259842520 0.8204275    NA         
#> 35  0.267716535 0.8229932    NA         
#> 36  0.275590551 0.8262549    NA         
#> 37  0.283464567 0.8301366    NA         
#> 38  0.291338583 0.8345534    NA         
#> 39  0.299212598 0.8394160    NA         
#> 40  0.307086614 0.8446342    NA         
#> 41  0.314960630 0.8501044    NA         
#> 42  0.322834646 0.8557374    NA         
#> 43  0.330708661 0.8614503    NA         
#> 44  0.338582677 0.8671684    NA         
#> 45  0.346456693 0.8728276    NA         
#> 46  0.354330709 0.8783750    NA         
#> 47  0.362204724 0.8837656    NA         
#> 48  0.370078740 0.8889747    NA         
#> 49  0.377952756 0.8939888    NA         
#> 50  0.385826772 0.8988056    NA         
#> 51  0.393700787 0.9034336    NA         
#> 52  0.401574803 0.9078908    NA         
#> 53  0.409448819 0.9122033    NA         
#> 54  0.417322835 0.9164019    NA         
#> 55  0.425196850 0.9205241    NA         
#> 56  0.433070866 0.9246097    NA         
#> 57  0.440944882 0.9286961    NA         
#> 58  0.448818898 0.9328188    NA         
#> 59  0.456692913 0.9370091    NA         
#> 60  0.464566929 0.9412926    NA         
#> 61  0.472440945 0.9456885    NA         
#> 62  0.480314961 0.9502097    NA         
#> 63  0.488188976 0.9548543    NA         
#> 64  0.496062992 0.9596141    NA         
#> 65  0.503937008 0.9644717    NA         
#> 66  0.511811024 0.9694005    NA         
#> 67  0.519685039 0.9743658    NA         
#> 68  0.527559055 0.9793254    NA         
#> 69  0.535433071 0.9842282    NA         
#> 70  0.543307087 0.9890195    NA         
#> 71  0.551181102 0.9936422    NA         
#> 72  0.559055118 0.9980370    NA         
#> 73  0.566929134 1.0021440    NA         
#> 74  0.574803150 1.0059046    NA         
#> 75  0.582677165 1.0092626    NA         
#> 76  0.590551181 1.0121586    NA         
#> 77  0.598425197 1.0145438    NA         
#> 78  0.606299213 1.0163853    NA         
#> 79  0.614173228 1.0176505    NA         
#> 80  0.622047244 1.0183153    NA         
#> 81  0.629921260 1.0183640    NA         
#> 82  0.637795276 1.0177899    NA         
#> 83  0.645669291 1.0165931    NA         
#> 84  0.653543307 1.0147732    NA         
#> 85  0.661417323 1.0123674    NA         
#> 86  0.669291339 1.0094059    NA         
#> 87  0.677165354 1.0059265    NA         
#> 88  0.685039370 1.0019745    NA         
#> 89  0.692913386 0.9976012    NA         
#> 90  0.700787402 0.9928630    NA         
#> 91  0.708661417 0.9878146    NA         
#> 92  0.716535433 0.9825265    NA         
#> 93  0.724409449 0.9770644    NA         
#> 94  0.732283465 0.9714912    NA         
#> 95  0.740157480 0.9658674    NA         
#> 96  0.748031496 0.9602491    NA         
#> 97  0.755905512 0.9546862    NA         
#> 98  0.763779528 0.9492224    NA         
#> 99  0.771653543 0.9438911    NA         
#> 100 0.779527559 0.9387104    NA         
#> 101 0.787401575 0.9336875    NA         
#> 102 0.795275591 0.9288152    NA         
#> 103 0.803149606 0.9240714    NA         
#> 104 0.811023622 0.9194188    NA         
#> 105 0.818897638 0.9148041    NA         
#> 106 0.826771654 0.9101555    NA         
#> 107 0.834645669 0.9053879    NA         
#> 108 0.842519685 0.9004040    NA         
#> 109 0.850393701 0.8950949    NA         
#> 110 0.858267717 0.8893421    NA         
#> 111 0.866141732 0.8830206    NA         
#> 112 0.874015748 0.8760011    NA         
#> 113 0.881889764 0.8681314    NA         
#> 114 0.889763780 0.8592935    NA         
#> 115 0.897637795 0.8493722    NA         
#> 116 0.905511811 0.8382561    NA         
#> 117 0.913385827 0.8258469    NA         
#> 118 0.921259843 0.8120623    NA         
#> 119 0.929133858 0.7968387    NA         
#> 120 0.937007874 0.7801143    NA         
#> 121 0.944881890 0.7618636    NA         
#> 122 0.952755906 0.7421187    NA         
#> 123 0.960629921 0.7209111    NA         
#> 124 0.968503937 0.6982973    NA         
#> 125 0.976377953 0.6743581    NA         
#> 126 0.984251969 0.6491972    NA         
#> 127 0.992125984 0.6229369    NA         
#> 128 1.000000000 0.5957027    NA